feat(orders): sell by piece, payment status and fractional package stock
- products.sell_by_piece; order_lines.pieces_quantity / sold_by_piece;
  orders.payment_status / paid_amount / paid_at; payments.order_id
- LineItem carries whole packages plus loose pieces; piece price is
  package price divided by units per package
- create order: piece keypad and draft persistence; order list and
  receipt: payment status and mark-as-paid
- package_stock helpers for fractional package stock on hand
- Supabase migrations 025-028; l10n strings for the above

+++ lib/core/notifications/notification_service.dart
@@ -128,9 +128,9 @@ class NotificationService {
           .gt('low_stock_threshold', 0);
 
       for (final product in List<Map<String, dynamic>>.from(result)) {
-        final stock = (product['stock_on_hand'] as num?)?.toInt() ?? 0;
+        final stock = (product['stock_on_hand'] as num?)?.toDouble() ?? 0;
         final threshold =
-            (product['low_stock_threshold'] as num?)?.toInt() ?? 0;
+            (product['low_stock_threshold'] as num?)?.toDouble() ?? 0;
         if (stock <= threshold) {
           await sendNotification(
             eventType: 'low_stock',
